Replace adjacency and topsort with AcyclicCheck
