+++ pipemimic/src/GraphTree.scala
